Replace package-version name in case of rc version
The deb packages for RC releases named with '~' as following
<version>~<rc_version> ex. 4.6~rc0  but the web-install script is
loooking for <version>.<rc_version> ex. 4.6.rc0

This commit adds the `is_rc_version` func to redefine the version in
case of rc version.

PS. We should change this implementation of deb packages with ~ since
it bother us in many places. IMO RC packages shouldn't be located in the
same folder as GA packages.